+++ praetorian_cli/handlers/list.py
@@ -16,8 +16,7 @@ def list(ctx):
     pass
 
 
-list_filter = {'jobs': 'updated', 'files': 'name', 'accounts': 'name', 'integrations': 'name', 'definitions': 'name',
-               'attributes': 'name'}
+list_filter = {'jobs': 'updated', 'files': 'name', 'accounts': 'name', 'integrations': 'name', 'definitions': 'name'}
 
 
 def attribute_filter(controller, key, offset, details, page):
@@ -37,10 +36,10 @@ def attribute_filter(controller, key, offset, details, page):
 @click.option('-attr', '--attribute', nargs=2, help='Filter by attribute name and value')
 @list_options('DNS')
 @page_options
-def assets(controller, filter, offset, details, page, attr):
+def assets(controller, filter, offset, details, page, attribute):
     """List assets"""
-    if attr:
-        attribute_filter(controller, f'#attribute#{attr[0]}#{attr[1]}#asset#{filter}', offset, details, page)
+    if attribute:
+        attribute_filter(controller, f'#attribute#{attribute[0]}#{attribute[1]}#asset#{filter}', offset, details, page)
         return
 
     paginate(controller, f'#asset#{filter}', 'assets', "", offset, details, page)
@@ -50,13 +49,29 @@ def assets(controller, filter, offset, details, page, attr):
 @list_options('name')
 @page_options
 @click.option('-attr', '--attribute', nargs=2, help='Filter by attribute name and value')
-def risks(controller, filter, offset, details, page, attr):
+def risks(controller, filter, offset, details, page, attribute):
     """List risks"""
-    if attr:
-        attribute_filter(controller, f'#attribute#{attr[0]}#{attr[1]}#risk#{filter}', details)
+    if attribute:
+        attribute_filter(controller, f'#attribute#{attribute[0]}#{attribute[1]}#risk#{filter}', details)
         return
 
     paginate(controller, f'#risk#{filter}', 'risks', "", offset, details, page)
+
+
+@list.command('attributes')
+@list_options('name')
+@page_options
+@click.option('-a', '--asset', help='Filter by asset key')
+@click.option('-r', '--risk', help='Filter by risk key')
+def attributes(controller, filter, offset, details, page, asset, risk):
+    """List attributes\n
+    You can only filter by one of the following : attribute name, asset or risk"""
+    key = f'#attribute#{filter}'
+    if asset:
+        key = f'source:{asset}'
+    elif risk:
+        key = f'source:{risk}'
+    paginate(controller, key, 'attributes', "", offset, details, page)
 
 
 def create_list_command(item_type, item_filter):