+++ praetorian_cli/handlers/unlink.py
@@ -11,25 +11,9 @@ def unlink(ctx):
     pass
 
 
-integration = ['slack', 'jira', 'amazon', 'azure', 'gcp', 'github', 'ns1', 'crowdstrike']
-
-
 @unlink.command('account')
 @click.argument('account_id')
 @cli_handler
 def unlink_account(controller, account_id):
     """ Unlink an account """
     controller.unlink(account_id, "")
-
-
-def unlink_integration(i):
-    @unlink.command(i, help=f"Unlink {i} integration")
-    @cli_handler
-    @click.option('-id', '--id', default="", help="Provide an id if there are multiple instances")
-    def command(controller, id):
-        """ Unlink an integration """
-        controller.unlink(i, id)
-
-
-for i in integration:
-    unlink_integration(i)